Code edits: implement condition classes and functions
Add DeprecatedCondition class
Add LocalError class
Add LocalWarning class
Add TypeError class
Add Error()
Add Warning()
Replace stop/warning() calls with throw()
Add check_assignment()
Add check_equal_weights()
Add check_integer()
Add check_numeric()
Add toString.character() with conjunction argument
Add scalar argument to check_logical()
Add domain checks for numeric MLControl slots
Set integer types on MLControl slots
Move condition code to conditions.R

+++ R/classes.R
@@ -255,7 +255,7 @@ setClass("MLControl",
 
 setClass("MLBootstrapControl",
   contains = "MLControl",
-  slots = c(samples = "numeric")
+  slots = c(samples = "integer")
 )
 
 
@@ -272,8 +272,8 @@ setClass("MLBootOptimismControl",
 setClass("MLCrossValidationControl",
   contains = "MLControl",
   slots = c(
-    folds = "numeric",
-    repeats = "numeric"
+    folds = "integer",
+    repeats = "integer"
   )
 )
 
@@ -290,7 +290,7 @@ setClass("MLCVOptimismControl",
 
 setClass("MLOOBControl",
   contains = "MLControl",
-  slots = c(samples = "numeric")
+  slots = c(samples = "integer")
 )
